Split out client into interface and implementation to allow mocking
